app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
cssselect2==0.8.0
dnspython==2.8.0
email-validator==2.3.0
faker==38.0.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
arq==0.25.0
redis==5.2.0
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.6.1
pydyf==0.11.0
pygments==2.19.2
pyjwt==2.10.1
pyotp==2.9.0
pyphen==0.17.2
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
sentry-sdk==2.44.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.49.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.4.0
telethon==1.36.0
pdf2image==1.17.0
openpyxl==3.1.5
pandas==2.2.0
